Move debug implementation inside module
This commit moves the debug signal into the `ptp_b` module, and adds another value input, which gets selected when the debug signal is asserted.

The selection between the signals is done with a ternary operator, and is assigned to the existing `internal_concat` signal.

The base logic for selecting which byte to send remains unchanged.